+++ pyi.py
@@ -568,6 +568,12 @@ class PyiVisitor(ast.NodeVisitor):
             self.generic_visit(node)
 
     def visit_AnnAssign(self, node: ast.AnnAssign) -> None:
+        if _is_name(node.target, "__all__") and not self.in_class.active:
+            with self.string_literals_allowed.enabled():
+                self.generic_visit(node)
+            if node.value is None:
+                self.error(node, Y035)
+            return
         self.generic_visit(node)
         if _is_TypeAlias(node.annotation):
             return
@@ -1124,3 +1130,4 @@ Y032 = (
     'Y032 Prefer "object" to "Any" for the second parameter in "{method_name}" methods'
 )
 Y033 = 'Y033 Do not use type comments in stubs (e.g. use "x: int" instead of "x = ... # type: int")'
+Y035 = 'Y035 "__all__" in a stub file should be identical to "__all__" at runtime'