+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Features:
 * introduce Y032 (prefer `object` to `Any` for the second argument in `__eq__` and
   `__ne__` methods).
 * introduce Y033 (always use annotations in stubs, rather than type comments).
+* introduce Y035 (`__all__` in a stub has the same semantics as at runtime).
 
 ## 22.1.0